Crypto assets: first-class crypto currencies and trades (#263) (#646)

+++ app/ui/core/src/commonTest/kotlin/com/moneymanager/ui/monzo/MonzoBalanceFixtureE2ETest.kt
@@ -129,7 +129,14 @@ class MonzoBalanceFixtureE2ETest : DbTest() {
                     }
                 assertEquals(expectedByName.keys.sorted(), actualByName.keys.sorted())
                 expectedByName.forEach { (accountName, expected) ->
-                    assertEquals(expected.balance, actualByName.getValue(accountName).balance.amount, accountName)
+                    assertEquals(
+                        expected.balance,
+                        actualByName
+                            .getValue(accountName)
+                            .balance.amount
+                            .toLong(),
+                        accountName,
+                    )
                 }
             }
         }
